Implement List<EntityType> fetchProductsFromIds(List<EntityType> emptyProducts) for ProductRepository, CustomerRepository, SalesInvoiceRepository, ServiceRepository, StockTransactionRepository. These function accepts a List of entities with all null fields but eId. Finds existing entity with this Id and creates a List of existing entities.
This allows client to send only id field of related entity when building relationship between two entities.

Example:

POST /customer/
{ "name": "someCustomer"}
which returns a Customer with "id": 0

so when creating a SalesInvoice we can
{
     "customer": {"id":0},
     "documentNumber": "AAA1212CCC",
}
which automatically puts the Customer with "name": "someCustomer" to customer field of created SalesInvoice entity. The benefit is we didn't need to provide all details of Customer entity.